edit alerts details data types

+++ Data/Data/ArtAmlAlertDetailView.cs
@@ -21,7 +21,7 @@ namespace Data.Data
         public string? AlertTypeCd { get; set; }
         public string? ScenarioName { get; set; }
         public long? ScenarioId { get; set; }
-        public byte? MoneyLaunderingRiskScore { get; set; }
+        public int? MoneyLaunderingRiskScore { get; set; }
         public DateTime? CreateDate { get; set; }
         public DateTime? RunDate { get; set; }
         public DateTime? CloseDate { get; set; }
@@ -29,6 +29,6 @@ namespace Data.Data
         public string? ReportCloseRsn { get; set; }
         public string? PoliticallyExposedPersonInd { get; set; }
         public string? EmployeeInd { get; set; }
-        public decimal? InvestigationDays { get; set; }
+        public int? InvestigationDays { get; set; }
     }
 }
